+++ android/src/main/java/com/pichillilorenzo/flutter_inappbrowser/chrome_custom_tabs/ChromeCustomTabsActivity.java
@@ -91,4 +91,5 @@ public class ChromeCustomTabsActivity extends Activity {
       InAppBrowserFlutterPlugin.channel.invokeMethod("onChromeSafariBrowserClosed", obj);
     }
   }
+
 }